-Put dictionary spaghetti into its own new class
